babystepping msg
on line 137 there was a typo "Only idiots try to babystep withoung homing"  and it always bothered me that someone would be called an idiot for pushing a button by mistake or maybe is still learning.  I've changed the msg to be more explanatory vs accusatory.  Hopefully this helps people learn instead of name call them for a mistake.